New Feature with Get Recipe
Adding the feature to get Recipes by their name even if the name is not finished

Adding the feature to get all Recipes from an author :)
